requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
apscheduler==3.10.4
python-multipart==0.0.6
